Corrección migración

+++ database/migrations/2020_03_24_135417_create_demographics_table.php
@@ -22,7 +22,7 @@ class CreateDemographicsTable extends Migration
             $table->string('department')->nullable();
             $table->string('nationality')->nullable();
 
-            $table->unsignedInteger('region_id');
+            $table->unsignedBigInteger('region_id');
             $table->string('region')->nullable();
             // $table->enum('region',['Arica y Parinacota',
             //                        'Tarapacá',
@@ -41,7 +41,7 @@ class CreateDemographicsTable extends Migration
             //                        'Región Metropolitana de Santiago',
             //                        'Región de Ñuble']);
 
-            $table->unsignedInteger('commune_id');
+            $table->unsignedBigInteger('commune_id');
             $table->string('commune')->nullable();
             // $table->enum('commune',['Arica', 'Camarones', 'Putre', 'General Lagos','Iquique', 'Alto Hospicio', 'Pozo Almonte', 'Camiña',
             //                         'Colchane', 'Huara', 'Pica','Antofagasta', 'Mejillones', 'Sierra Gorda', 'Taltal', 'Calama', 'Ollagüe',
